feat: set items as esm
