+++ .husky/pre-commit.yaml
@@ -1,0 +1,2 @@
+pnpm format
+pnpm lint
