+++ apps/web/.eslintrc.mjs
@@ -1,0 +1,18 @@
+/** @type {import("eslint").Linter.Config} */
+// module.exports = {
+//   root: true,
+//   extends: ["@tripie/eslint-config/next.js"],
+//   parser: "@typescript-eslint/parser",
+//   parserOptions: {
+//     project: true,
+//   },
+// };
+
+export default {
+  root: true,
+  extends: ["@tripie/eslint-config/next.js"],
+  parser: "@typescript-eslint/parser",
+  parserOptions: {
+    project: true,
+  },
+};
