+++ apps/web/app/page.tsx
@@ -1,10 +1,8 @@
-import { MyButton } from "@tripie/design-system";
-
 export default function Home() {
   return (
     <div>
-      this is home
-      <MyButton>Click MyButton!</MyButton>
+      {/* this is home */}
+      {/* <MyButton>Click MyButton!</MyButton> */}
     </div>
   );
 }
