+++ apps/web/next.config.mjs
@@ -1,0 +1,17 @@
+/**
+ * https://nextjs.org/docs/app/building-your-application/styling/sass#customizing-sass-options
+ */
+
+import path from "path";
+import { fileURLToPath } from "url";
+
+const __dirname = fileURLToPath(new URL(".", import.meta.url));
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  sassOptions: {
+    includePaths: [path.join(__dirname, "styles")],
+  },
+};
+
+export default nextConfig;
