+++ apps/web/package.json
@@ -26,6 +26,5 @@
     "eslint-config-next": "15.0.0-rc.0",
     "node-sass": "^9.0.0",
     "typescript": "^5"
-  },
-  "type": "module"
+  }
 }
